Changed some int to size_t

+++ src/Data.cpp
@@ -11,7 +11,10 @@ ezc3d::DataNS::Data::Data()
 ezc3d::DataNS::Data::Data(ezc3d::c3d &file)
 {
     // Firstly read a dummy value just prior to the data so it moves the pointer to the right place
-    file.readInt(ezc3d::DATA_TYPE::BYTE, 256*ezc3d::DATA_TYPE::WORD*(file.header().parametersAddress()-1) + 256*ezc3d::DATA_TYPE::WORD*file.parameters().nbParamBlock() - ezc3d::DATA_TYPE::BYTE, std::ios::beg); // "- BYTE" so it is just prior
+    file.readInt(ezc3d::DATA_TYPE::BYTE,
+                 static_cast<int>(256*ezc3d::DATA_TYPE::WORD*(file.header().parametersAddress()-1) +
+                 256*ezc3d::DATA_TYPE::WORD*file.parameters().nbParamBlock() -
+                 ezc3d::DATA_TYPE::BYTE), std::ios::beg); // "- BYTE" so it is just prior
 
     // Initialize some variables
     if (file.header().nbFrames()>0)
